functions/requirements.txt
azure-functions==1.21.3
azure-storage-blob==12.24.0
azure-search-documents==11.5.2
azure-identity==1.19.0
tiktoken==0.8.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0